move utils to upper level
