+++ src/cli/env.js
@@ -1,5 +1,8 @@
+import { env } from 'node:process';
+
 const parseEnv = () => {
-    // Write your code here 
+
+console.log(env);
 };
 
 parseEnv();
