+++ src/fs/list.js
@@ -1,11 +1,11 @@
-import { checkIfPathExists, errorText, getFilesList, getFilesFolderPath } from "./utils.js";
+import { checkIfPathExists, errorText, getFilesList, getFilesFolderPath } from "../utils.js";
 
 const list = async () => {
-  if (!checkIfPathExists(getFilesFolderPath())) {
+  if (!checkIfPathExists(getFilesFolderPath('fs'))) {
     throw new Error(errorText);
   }
 
-  console.log(await getFilesList());
+  console.log(await getFilesList('fs'));
 };
 
 await list();
